requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.7
Django==4.1.13
django-debug-toolbar==4.2.0
django-filter==23.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.6.1
fake-useragent==1.4.0
filelock==3.13.1
fonttools==4.53.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==2.0.0
packaging==24.1
pillow==10.3.0
pipenv==2023.11.15
platformdirs==4.1.0
pymongo==3.12.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.2.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.25.0
